add preferred validators to assisted flow
