+++ hooks/useAssistedUnstake.ts
@@ -1,19 +1,18 @@
 import { useState } from 'react';
 import { useConnection, useWallet } from '@solana/wallet-adapter-react';
-// import * as splStakePool from '@solana/spl-stake-pool/dist/index.cjs';
 import { TransactionMessage, VersionedTransaction, ComputeBudgetProgram, Transaction, PublicKey } from '@solana/web3.js';
 import { COMPUTE_UNIT_LIMIT_FOR_STAKE_OPERATIONS, JITO_STAKE_POOL_ADDRESS } from '../constants';
 import toast from 'react-hot-toast';
 import * as solanaStakePool from '@solana/spl-stake-pool';
 import { useNetwork } from '../components/NetworkProvider';
 import { WalletAdapterNetwork } from '@solana/wallet-adapter-base';
+import { PreferredWithdraw, getPreferredValidatorsApiUrl } from '../utils/preferredValidators';
 
 /**
  * Parameters for unstaking JitoSOL
  */
 export interface UnstakeParams {
     useReserve: boolean;
-    voteAccountAddress?: PublicKey;
     stakeReceiver?: PublicKey;
 }
 
@@ -61,6 +60,35 @@ export const useAssistedUnstake = () => {
                 return false;
             }
 
+            // Fetch preferred validators from API
+            let voteAccountAddress: PublicKey | undefined;
+
+            try {
+                toast.loading('Fetching preferred validators...', { id: toastId });
+                const apiUrl = getPreferredValidatorsApiUrl(network);
+
+                const response = await fetch(apiUrl, {
+                    method: 'GET',
+                    headers: {
+                        'Content-Type': 'application/json',
+                    },
+                });
+
+                if (response.ok) {
+                    const preferredValidators: PreferredWithdraw[] = await response.json();
+
+                    if (Array.isArray(preferredValidators) && preferredValidators.length > 0) {
+                        // Take the first validator from the list
+                        const selectedValidator = preferredValidators[0];
+                        voteAccountAddress = new PublicKey(selectedValidator.vote_account);
+                        console.log(`Using preferred validator: ${voteAccountAddress.toBase58()} (rank ${selectedValidator.rank})`);
+                    }
+                }
+            } catch (apiError) {
+                console.warn('Failed to fetch preferred validators, will use default:', apiError);
+                // Continue without preferred validator - SPL library will auto-select
+            }
+
             // Call withdrawStake with parameters based on the unstake options
             const withdrawStakeArgs = [
                 connection,                           // connection
@@ -70,15 +98,15 @@ export const useAssistedUnstake = () => {
                 unstakeParams.useReserve              // useReserve
             ];
 
-            // Add optional parameters if provided
-            if (unstakeParams.voteAccountAddress) {
-                withdrawStakeArgs.push(unstakeParams.voteAccountAddress); // voteAccountAddress (6th arg)
+            // Add voteAccountAddress if we have one (from API)
+            if (voteAccountAddress) {
+                withdrawStakeArgs.push(voteAccountAddress); // voteAccountAddress (6th arg)
             }
 
             if (unstakeParams.stakeReceiver) {
                 // withdrawStake uses positional args. If stakeReceiver (7th) is provided
                 // but voteAccountAddress (6th) is not, we must provide a placeholder for the 6th arg.
-                if (!unstakeParams.voteAccountAddress) {
+                if (!voteAccountAddress) {
                     withdrawStakeArgs.push(undefined as any);
                 }
                 withdrawStakeArgs.push(unstakeParams.stakeReceiver); // stakeReceiver (7th arg)
